requirements.txt
PyRSS2Gen
redis
python-dotenv